DBZ-7223 Move sink tests into a separate package
