+++ debezium-connector-mongodb/src/test/java/io/debezium/connector/mongodb/sink/SinkConnectorShardedClusterIT.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.connector.mongodb;
+package io.debezium.connector.mongodb.sink;
 
 import static io.debezium.connector.mongodb.TestHelper.cleanDatabase;
 import static org.hamcrest.CoreMatchers.is;
@@ -15,12 +15,16 @@ import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
+import io.debezium.connector.mongodb.AbstractShardedMongoConnectorIT;
 import io.debezium.connector.mongodb.junit.MongoDbDatabaseProvider;
+import io.debezium.connector.mongodb.junit.MongoDbDatabaseVersionResolver;
+import io.debezium.connector.mongodb.junit.MongoDbPlatform;
 import io.debezium.testing.testcontainers.MongoDbDeployment;
+import io.debezium.testing.testcontainers.MongoDbShardedCluster;
 import io.debezium.testing.testcontainers.testhelper.TestInfrastructureHelper;
 import io.debezium.testing.testcontainers.util.DockerUtils;
 
-public class SinkConnectorReplicaSetIT extends AbstractMongoConnectorIT implements SinkConnectorIT {
+public class SinkConnectorShardedClusterIT extends AbstractShardedMongoConnectorIT implements SinkConnectorIT {
 
     protected static MongoDbDeployment mongo;
 
@@ -34,14 +38,22 @@ public class SinkConnectorReplicaSetIT extends AbstractMongoConnectorIT implemen
         Assume.assumeThat("Skipping DebeziumMongoDbConnectorResourceIT tests when assembly profile is not active!",
                 System.getProperty("isAssemblyProfileActive", "false"),
                 is("true"));
+        Assume.assumeTrue(MongoDbDatabaseVersionResolver.getPlatform().equals(MongoDbPlatform.MONGODB_DOCKER));
         DockerUtils.enableFakeDnsIfRequired();
-        mongo = MongoDbDatabaseProvider.externalOrDockerReplicaSet(TestInfrastructureHelper.getNetwork());
+        mongo = MongoDbDatabaseProvider.mongoDbShardedCluster(TestInfrastructureHelper.getNetwork());
         mongo.start();
     }
 
     @Before
     public void beforeEach() {
         sendSourceData();
+
+        var database = shardedDatabase();
+        var shardedCluster = (MongoDbShardedCluster) mongo;
+        shardedCluster.enableSharding(database);
+        shardedCollections().forEach((collection, key) -> {
+            shardedCluster.shardCollection(database, collection, key);
+        });
     }
 
     @After
@@ -56,8 +68,7 @@ public class SinkConnectorReplicaSetIT extends AbstractMongoConnectorIT implemen
     }
 
     @Test
-    public void testSinkConnectorWritesRecordsToReplicaSet() {
+    public void testSinkConnectorWritesRecordsToShardedCluster() {
         checkSinkConnectorWritesRecords();
     }
-
 }
